requirements.txt
pandas>=1.3.0
pyspark>=3.2.0
pyodbc>=4.0.30
JayDeBeApi>=1.2.3
scikit-learn>=1.0.0
pytest>=6.0.0
python-dotenv>=0.19.0
numpy>=1.20.0 